Updated to have the moves become deprecations instead. Will submit a new PR for planned_3.0.0_changes to resolve the deprecation.
