+++ jave-core/src/main/java/ws/schild/process/ProcessWrapper.java
@@ -16,7 +16,7 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package ws.schild.jave;
+package ws.schild.process;
 
 import java.io.IOException;
 import java.io.InputStream;
@@ -26,15 +26,13 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 /**
- * An ffmpeg process wrapper.
+ * A ffmpeg process wrapper.
  *
  * @author Carlo Pelliccia
- * @deprecated  As of 3.0.0. Use {@link ws.schild.process.ProcessWrapper} instead.
  */
-@Deprecated
-public class FFMPEGExecutor {
+public class ProcessWrapper {
 
-    private static final Logger LOG = LoggerFactory.getLogger(FFMPEGExecutor.class);
+    private static final Logger LOG = LoggerFactory.getLogger(ProcessWrapper.class);
 
     /**
      * The path of the ffmpeg executable.
@@ -77,7 +75,7 @@ public class FFMPEGExecutor {
      *
      * @param ffmpegExecutablePath The path of the ffmpeg executable.
      */
-    public FFMPEGExecutor(String ffmpegExecutablePath) {
+    public ProcessWrapper(String ffmpegExecutablePath) {
         this.ffmpegExecutablePath = ffmpegExecutablePath;
     }
